Use new location for GTFS RT models
