+++ apps/_shared/ts/types/gtfs-rt/entity.ts
@@ -1,0 +1,7 @@
+import { TripUpdate } from './gtfs-rt'
+
+export interface Response_GTFS_RT_Entity {
+    Id: string,
+    IsDeleted: boolean,
+    TripUpdate?: TripUpdate
+}
